+++ apps/builder/app/builder/features/style-panel/sections/shared/value-text.tsx
@@ -1,8 +1,9 @@
 import { styled, Text } from "@webstudio-is/design-system";
 import type { StyleValue } from "@webstudio-is/css-engine";
-import { useMemo, type ComponentProps } from "react";
+import { useEffect, useMemo, type ComponentProps } from "react";
 import { theme } from "@webstudio-is/design-system";
 import { toValue } from "@webstudio-is/css-engine";
+import { scrollByPointer } from "../../shared/scroll-by-pointer";
 
 const Container = styled("button", {
   // fit-content is not needed for the "button" element,
@@ -61,21 +62,13 @@ const Container = styled("button", {
 export const ValueText = ({
   value,
   source,
-  truncate = false,
   ...rest
-}: { value: StyleValue; truncate?: boolean } & Omit<
-  ComponentProps<typeof Container>,
-  "value"
->) => {
+}: { value: StyleValue } & Omit<ComponentProps<typeof Container>, "value">) => {
   const children = useMemo(() => {
     if (value.type === "unit") {
       // we want to show "0" rather than "0px" for default values for cleaner UI
       if (source === "default" && value.unit === "px" && value.value === 0) {
-        return (
-          <Text truncate={truncate} variant="spaceSectionValueText">
-            {value.value}
-          </Text>
-        );
+        return <Text variant="spaceSectionValueText">{value.value}</Text>;
       }
 
       /**
@@ -85,7 +78,7 @@ export const ValueText = ({
 
       return (
         <>
-          <Text truncate={truncate} variant="spaceSectionValueText">
+          <Text variant="spaceSectionValueText">
             {value.value}
             <Text
               variant="spaceSectionValueText"
@@ -102,22 +95,20 @@ export const ValueText = ({
     }
 
     if (value.type === "var") {
-      return (
-        <Text truncate={truncate} variant="spaceSectionValueText">
-          {value.value}
-        </Text>
-      );
+      return <Text variant="spaceSectionValueText">{value.value}</Text>;
     }
 
-    return (
-      <Text truncate={truncate} variant="spaceSectionValueText">
-        {toValue(value)}
-      </Text>
-    );
-  }, [value, source, truncate]);
+    return <Text variant="spaceSectionValueText">{toValue(value)}</Text>;
+  }, [value, source]);
+
+  const { abort, ...autoScrollProps } = useMemo(scrollByPointer, []);
+
+  useEffect(() => {
+    return () => abort("unmount");
+  }, [abort]);
 
   return (
-    <Container source={source} {...rest} tabIndex={-1}>
+    <Container source={source} {...rest} {...autoScrollProps} tabIndex={-1}>
       {children}
     </Container>
   );
