+++ apps/builder/app/builder/features/style-panel/shared/css-value-input/css-value-input.tsx
@@ -58,6 +58,7 @@ import {
   ValueEditorDialog,
 } from "./value-editor-dialog";
 import { useEffectEvent } from "~/shared/hook-utils/effect-event";
+import { scrollByPointer } from "../scroll-by-pointer";
 
 // We need to enable scrub on properties that can have numeric value.
 const canBeNumber = (property: CssProperty, value: CssValueInputValue) => {
@@ -343,93 +344,6 @@ const itemToString = (item: CssValueInputValue | null) => {
     return String(item.value);
   }
   return toValue(item);
-};
-
-const scrollAhead = ({ target, clientX }: MouseEvent) => {
-  const element = target as HTMLInputElement;
-
-  if (element.scrollWidth === element.clientWidth) {
-    // Nothing to scroll.
-    return false;
-  }
-  const inputRect = element.getBoundingClientRect();
-
-  // Calculate the relative x position of the mouse within the input element
-  const relativeMouseX = clientX - inputRect.x;
-
-  // Calculate the percentage position (0% at the beginning, 100% at the end)
-  const inputWidth = inputRect.width;
-  const mousePercentageX = Math.ceil((relativeMouseX / inputWidth) * 100);
-
-  // Apply acceleration based on the relative position of the mouse
-  // Closer to the beginning (-20%), closer to the end (+20%)
-  const accelerationFactor = (mousePercentageX - 50) / 50;
-  const adjustedMousePercentageX = Math.min(
-    Math.max(mousePercentageX + accelerationFactor * 20, 0),
-    100
-  );
-
-  // Calculate the scroll position corresponding to the adjusted percentage
-  const scrollPosition =
-    (adjustedMousePercentageX / 100) *
-    (element.scrollWidth - element.clientWidth);
-
-  // Scroll the input element
-  element.scroll({ left: scrollPosition });
-  return true;
-};
-
-const getAutoScrollProps = () => {
-  let abortController = new AbortController();
-
-  const abort = (reason: string) => {
-    abortController.abort(reason);
-  };
-
-  return {
-    abort,
-    onMouseOver(event: MouseEvent) {
-      if (event.target === document.activeElement) {
-        abort("focused");
-        return;
-      }
-      if (scrollAhead(event) === false) {
-        return;
-      }
-
-      abortController = new AbortController();
-      event.target?.addEventListener(
-        "mousemove",
-        (event) => {
-          if (event.target === document.activeElement) {
-            abort("focused");
-            return;
-          }
-          requestAnimationFrame(() => {
-            scrollAhead(event as MouseEvent);
-          });
-        },
-        {
-          signal: abortController.signal,
-          passive: true,
-        }
-      );
-    },
-    onMouseOut(event: MouseEvent) {
-      if (event.target === document.activeElement) {
-        abort("focused");
-        return;
-      }
-      (event.target as HTMLInputElement).scroll({
-        left: 0,
-        behavior: "smooth",
-      });
-      abort("mouseout");
-    },
-    onFocus() {
-      abort("focus");
-    },
-  };
 };
 
 const Description = styled(Box, { width: theme.spacing[27] });
@@ -880,7 +794,7 @@ export const CssValueInput = ({
   };
 
   const { abort, ...autoScrollProps } = useMemo(() => {
-    return getAutoScrollProps();
+    return scrollByPointer();
   }, []);
 
   useEffect(() => {
